+++ ccr/pom.xml
@@ -16,7 +16,6 @@
 
 	<dependencies>
 
-
 		<!-- JSF 2.2 - ok -->
 		<dependency>
 			<groupId>com.sun.faces</groupId>
@@ -30,8 +29,6 @@
 			<version>2.2.13</version>
 		</dependency>
 
-
-
 		<!-- Jersey 2.23.2 - ok -->
 		<dependency>
 			<groupId>org.glassfish.jersey.containers</groupId>
@@ -44,8 +41,6 @@
 			<version>2.23.2</version>
 		</dependency>
 
-
-
 		<!-- CDI for JSF 2.2 - ok -->
 		<dependency>
 			<groupId>javax.enterprise</groupId>
@@ -53,8 +48,6 @@
 			<version>2.0</version>
 			<scope>provided</scope>
 		</dependency>
-
-
 
 		<!-- Servlet e JSP - ok -->
 		<dependency>
@@ -221,9 +214,6 @@
 			<version>1.2.1</version>
 		</dependency>
 
-
-
-
 		<dependency>
 			<groupId>org.omnifaces</groupId>
 			<artifactId>omnifaces</artifactId>
@@ -244,13 +234,18 @@
 			<version>2.2.0</version>
 		</dependency>
 
-
-
 		<!-- JUNIT -->
 		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
 			<version>4.5</version>
+			<scope>test</scope>
+		</dependency>
+
+		<dependency>
+			<groupId>org.hamcrest</groupId>
+			<artifactId>hamcrest-core</artifactId>
+			<version>1.3</version>
 			<scope>test</scope>
 		</dependency>
