+++ ccr/src/main/resources/messages.properties
@@ -88,13 +88,13 @@ rodents = Rodents
 ferrets = Ferrets
 horses = Horses
 crickets = Crickets
-addPet = Add Certificado
+addCertificate = Add Certificado
 itemToAdd = New
 breed = Breed
 age = Age
 pedigree = Pedigree
 mating = Mating
-petType = Pet type
+certType = Certificate Type
 charsRemaining = characters remaining.
 invalidSizeMessage = Size limit exceeded!
 invalidFileMessage = File type is invalid!
