+++ ccr/src/main/resources/messages_en_US.properties
@@ -88,13 +88,13 @@ rodents = Rodents
 ferrets = Ferrets
 horses = Horses
 crickets = Crickets
-addPet = Add Certificate
+addCertificate = Add Certificate
 itemToAdd = New
 breed = Breed
 age = Age
 pedigree = Pedigree
 mating = Mating
-petType = Pet type
+certType = Certificate type
 charsRemaining = characters remaining.
 invalidSizeMessage = Size limit exceeded!
 invalidFileMessage = File type is invalid!
