+++ ccr/src/main/resources/messages_es_ES.properties
@@ -86,13 +86,13 @@ rodents = Roedores
 ferrets = Hurones
 horses = Caballos
 crickets = Grillos
-addPet = A\u00f1adir Certificado
+addCertificate = A\u00f1adir Certificado
 itemToAdd = Nuevo
 breed = Raza
 age = Edad
 pedigree = \u00c1rbol geneal\u00f3gico
 mating = Apareamiento
-petType = tipo de mascota
+certType = tipo de certificado
 charsRemaining = caracteres restantes.
 invalidSizeMessage = \u00a1L\u00edmite de tama\u00f1o supera!
 invalidFileMessage = Tipo de archivo no es v\u00e1lido!
