+++ ccr/src/main/resources/messages_pt_BR.properties
@@ -86,13 +86,13 @@ rodents = Roedores
 ferrets = Fur\u00f5es
 horses = Cavalos
 crickets = Grilos
-addPet = Adicionar Certificado
+addCertificate = Adicionar Certificado
 itemToAdd = Novo
 breed = Ra\u00e7a
 age = Idade
 pedigree = Pedigree (\u00c1rvore Geneal\u00f3gica)
 mating = Acasalamento
-petType = Tipo de animal de estima\u00e7\u00e3o
+certType = Tipo de certificado
 charsRemaining = caracteres restantes.
 invalidSizeMessage = Tamanho limite de arquivo excedido!
 invalidFileMessage = Tipo de arquivo \u00e9 inv\u00e1lido!
