+++ ccr/src/main/webapp/META-INF/context.xml
@@ -1,4 +1,28 @@
 <?xml version="1.0" encoding="UTF-8"?>
+
+<!--
 <Context antiJARLocking="true" path="/ccr">
 	<Realm className="org.apache.catalina.realm.MemoryRealm" />
+
+	<Resource name="jdbc/ccr" auth="Container"
+		type="javax.sql.DataSource" username="root" password="JUJR*mysql21"
+		driverClassName="com.mysql.jdbc.Driver"
+		url="jdbc:mysql://localhost:3306/ccr?useUnicode=true&amp;characterEncoding=utf8"
+		maxTotal="15" maxIdle="3" defaultTransactionIsolation="READ_COMMITTED"
+		validationQuery="Select 1" />
+
 </Context>
+-->
+
+<Context antiJARLocking="true" path="/ccr">
+	<Realm className="org.apache.catalina.realm.MemoryRealm" />
+
+	<Resource name="jdbc/ccr" auth="Container"
+		type="javax.sql.DataSource" username="sa4mysql@ccr21mysqlsrv" password="DnLKWNxczX6tEbBmCTpk"
+		driverClassName="com.mysql.jdbc.Driver"
+		url="jdbc:mysql://ccr21mysqlsrv.mysql.database.azure.com:3306/ccr?useUnicode=true&amp;characterEncoding=utf8&amp;useSSL=true&amp;=requireSSL=false"
+		maxTotal="15" maxIdle="3" defaultTransactionIsolation="READ_COMMITTED"
+		validationQuery="Select 1" />
+
+</Context>
+
