+++ ccr/src/main/webapp/WEB-INF/faces-config.xml
@@ -29,8 +29,8 @@
 		<managed-bean-scope>session</managed-bean-scope>
 	</managed-bean>
 	<managed-bean>
-		<managed-bean-name>certMB</managed-bean-name>
-		<managed-bean-class>ie.lyit.ccr.control.CertificateMB</managed-bean-class>
+		<managed-bean-name>courseMB</managed-bean-name>
+		<managed-bean-class>ie.lyit.ccr.control.CourseMB</managed-bean-class>
 		<managed-bean-scope>session</managed-bean-scope>
 	</managed-bean>
 	<managed-bean>
