+++ ccr/src/main/webapp/WEB-INF/web.xml
@@ -88,43 +88,52 @@
 		<welcome-file>/index.jsp</welcome-file>
 	</welcome-file-list>
 
+	<resource-ref>
+		<description>CCR JDBC Data Source</description>
+		<res-ref-name>jdbc/ccr</res-ref-name>
+		<res-type>javax.sql.DataSource</res-type>
+		<res-auth>Container</res-auth>
+	</resource-ref>
+	
+	<!-- 
+	
 	<security-constraint>
-		<web-resource-collection>
-			<web-resource-name>CCR Web Resources</web-resource-name>
-			<url-pattern>/*</url-pattern>
-			<url-pattern>*.xhtml</url-pattern>
-			<url-pattern>*.faces</url-pattern>
-			<url-pattern>*.jsf</url-pattern>
-			<url-pattern>*.jsp</url-pattern>
-		</web-resource-collection>
-		<auth-constraint>
-			<role-name>ccradmin</role-name>
-			<role-name>ccruser</role-name>
-		</auth-constraint>
+<web-resource-collection>
+<web-resource-name>CCR Web Resources</web-resource-name>
+<url-pattern>/*</url-pattern>
+<url-pattern>*.xhtml</url-pattern>
+<url-pattern>*.faces</url-pattern>
+<url-pattern>*.jsf</url-pattern>
+<url-pattern>*.jsp</url-pattern>
+</web-resource-collection>
+<auth-constraint>
+<role-name>ccradmin</role-name>
+<role-name>ccruser</role-name>
+</auth-constraint>
 
-		<user-data-constraint>
-			<!-- transport-guarantee can be CONFIDENTIAL, INTEGRAL, or NONE -->
-			<transport-guarantee>NONE</transport-guarantee>
-		</user-data-constraint>
+<user-data-constraint>
+<transport-guarantee>NONE</transport-guarantee>
+</user-data-constraint>
 
-	</security-constraint>
+</security-constraint>
 
-	<login-config>
-		<auth-method>FORM</auth-method>
-		<realm-name>ccr</realm-name>
-		<form-login-config>
-			<form-login-page>/login.jsp</form-login-page>
-			<form-error-page>/loginerror.jsp</form-error-page>
-		</form-login-config>
-	</login-config>
+<login-config>
+<auth-method>FORM</auth-method>
+<realm-name>ccr</realm-name>
+<form-login-config>
+<form-login-page>/login.jsp</form-login-page>
+<form-error-page>/loginerror.jsp</form-error-page>
+</form-login-config>
+</login-config>
 
-	<security-role>
-		<role-name>ccruser</role-name>
-	</security-role>
-	<security-role>
-		<role-name>ccradmin</role-name>
-	</security-role>
-
+<security-role>
+<role-name>ccruser</role-name>
+</security-role>
+<security-role>
+<role-name>ccradmin</role-name>
+</security-role>
+	
+	 -->
 
 
 </web-app>
